requirements.txt
Flask==3.1.0
folium==0.19.4
geopandas==1.0.1
jupyter==1.1.1
jupyterlab==4.3.4
kaleido==0.2.1
numpy==2.2.2
pandas==2.2.3
plotly==5.24.1
polars==1.21.0
psycopg2-binary==2.9.10
pyarrow==19.0.0
spacy==3.8.4
python-dotenv==1.0.1
SQLAlchemy==2.0.37
